requirements.txt
langchain
streamlit
openai
pdfminer.six
sentence-transformers
faiss-cpu